Rename package
https://github.com/VEuPathDB/service-sequence-retrieval/pull/1#discussion_r866204302
