+++ src/main/java/org/veupathdb/service/sr/controller/SequenceController.java
@@ -1,16 +1,16 @@
-package org.veupathdb.service.sequence.controller;
+package org.veupathdb.service.sr.controller;
 
 import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.Request;
 import jakarta.ws.rs.core.StreamingOutput;
 
 import org.veupathdb.lib.container.jaxrs.server.annotations.Authenticated;
-import org.veupathdb.service.sequence.generated.model.SequencePostRequest;
-import org.veupathdb.service.sequence.generated.resources.SequenceSequenceType;
+import org.veupathdb.service.sr.generated.model.SequencePostRequest;
+import org.veupathdb.service.sr.generated.resources.SequenceSequenceType;
 
-import org.veupathdb.service.sequence.util.StreamSequences;
-import org.veupathdb.service.sequence.util.Validate;
-import org.veupathdb.service.sequence.service.Sequences;
+import org.veupathdb.service.sr.util.StreamSequences;
+import org.veupathdb.service.sr.util.Validate;
+import org.veupathdb.service.sr.service.Sequences;
 
 public class SequenceController implements SequenceSequenceType {
